==> scRNAseq_OSCA_Script/06_differential_abundance_analysis.Rmd <==
---
title: "06_differential_abundance_analysis"
author: ""
date: "2022/10/24"
output: 
  html_document: 
    toc: yes
    toc_depth: 5
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE)
```


## Load library
```{r}
library(SingleCellExperiment)
library(BiocParallel)
library(scater)
library(scuttle)
library(future)
library(ggplot2)
library(ggprism)
library(edgeR)
library(limma)
library(DESeq2)
library(EnhancedVolcano)
library(ComplexHeatmap)
library(dplyr)
library(ggpubr)
```

## Data
```{r}
# Parameters
this.path <- "../Results/06_differential_abundance_analysis"
bp.params <- MulticoreParam(workers = 20)
set.seed(1234)
dir.create(this.path)

# Import sce object before data integration
uncorrected <- readRDS("../Results/02_normalization_and_dimensionality_reduction_clustering/DataIntegration_uncorrected_PCA_TSNE.RDS")

# Import sce object after data integration by fastMNN
corrected <- readRDS("../Results/04_cell_type_annotation/DataIntegration_MNN_out_clustering_dimensionreduction_annotationbySingleR.RDS")

# Add cell types onto uncorrected data
uncorrected$cluster_id <- corrected$pruned.labels

# Remove not annotated cells
sce <- uncorrected[,!is.na(uncorrected$cluster_id)]
```

## Differential abundance between conditions

```{r}
n_cells <- table(sce$cluster_id, sce$Sample) 
freqs <- prop.table(n_cells, margin = 2)
abundances <- unclass(n_cells) 

coldata <- data.frame(colData(sce))
coldata <- coldata[!duplicated(coldata$Sample),]


df <- as.data.frame(t(freqs))
colnames(df)[1:3] <- c("Sample", "cluster_id", "frequency")
df <- left_join(df, coldata[,c("Sample", "cluster")])

ggplot(df, aes(x = Sample, y = frequency, fill = cluster_id)) +
    geom_bar(stat = "identity", position = "fill") + 
    facet_wrap(~ cluster, scales = "free_x", ncol = 4) +
    scale_fill_manual(values = c(as.character(see::material_colors()),
                      as.character(see::see_colors()))) +
    theme_classic() + 
  theme(axis.text.x = element_text(angle = 45, hjust = 1))
ggsave(paste0(this.path, "/cell_type_abundance_sample_barplot.pdf"), width = 9, height = 5)


ggplot(df, aes(x = cluster, y = frequency, fill = cluster_id)) +
    geom_bar(stat = "identity", position = "fill") + 
    # facet_wrap(~ diagnosis_id, scales = "free_x") +
    scale_fill_manual(values = c(as.character(see::material_colors()),
                      as.character(see::see_colors()))) +
    theme_classic() + 
  theme(axis.text.x = element_text(angle = 45, hjust = 1))
ggsave(paste0(this.path, "/cell_type_abundance_cluster_barplot.pdf"), width = 5, height = 6.5)

ggplot(df, aes(x = cluster, y = frequency, color = cluster)) +
    geom_boxplot(outlier.colour = NA) +  geom_jitter() +
    facet_wrap(~ cluster_id, scales = "free_y", ncol = 4) +
    theme_classic() + 
    scale_colour_prism(palette = "prism_light")
ggsave(paste0(this.path, "/cell_type_abundance_boxplot.pdf"), width = 9, height = 10)

# Attaching some column metadata.
extra.info <- colData(sce)[match(colnames(abundances), sce$Sample),]

# Creates a DGEList object from a table of counts
y.ab <- DGEList(abundances, samples=extra.info)

# Filter out low-abundance labels 
keep <- filterByExpr(y.ab, group=y.ab$samples$cluster)
y.ab <- y.ab[keep,]

# Formulate the design matrix 
design <- model.matrix(~factor(cluster), y.ab$samples)

# Estimate the NB dispersion for each cluster
y.ab <- estimateDisp(y.ab, design, trend="none")

# QL dispersion
fit.ab <- glmQLFit(y.ab, design, robust=TRUE, abundance.trend=FALSE)

# Test for differences in abundance
res.1 <- glmQLFTest(fit.ab, contrast = c(1,-1,0))
res.2 <- glmQLFTest(fit.ab, contrast = c(1,0,-1))
res.3 <- glmQLFTest(fit.ab, contrast = c(0,1,-1))

res_df.1 <- topTags(res.1, n = Inf)$table %>% round(2) %>% rownames_to_column(var = "cell_type")
res_df.2 <- topTags(res.2, n = Inf)$table %>% round(2) %>% rownames_to_column(var = "cell_type")
res_df.3 <- topTags(res.3, n = Inf)$table %>% round(2) %>% rownames_to_column(var = "cell_type")

openxlsx::write.xlsx(res_df.1, file = paste0(this.path, "/edgeR_differential_abundance_results_SHHD_vs_1.xlsx"))
openxlsx::write.xlsx(res_df.2, file = paste0(this.path, "/edgeR_differential_abundance_results_SHHD_vs_2.xlsx"))
openxlsx::write.xlsx(res_df.3, file = paste0(this.path, "/edgeR_differential_abundance_results_1_vs_2.xlsx"))
```



```{r}
abundances <- table(sce$cluster_id, sce$Sample) 
abundances <- unclass(abundances) 
head(abundances)

# Attaching some column metadata
extra.info <- colData(sce)[match(colnames(abundances), sce$Sample),]

# Creates a DGEList object from a table of counts
y.ab <- DGEList(abundances, samples=extra.info)
y.ab

# Determine which cell types have sufficiently large counts to be retained in a statistical analysis
keep <- filterByExpr(y.ab, group=y.ab$samples$cluster)
y.ab <- y.ab[keep,]
summary(keep)

# Creates a design (or model) matrix
design <- model.matrix(~factor(cluster), y.ab$samples)

# Estimate the NB dispersion for each cluster
y.ab <- estimateDisp(y.ab, design, trend="none")
summary(y.ab$common.dispersion)

# Plot the genewise biological coefficient of variation (BCV) against gene abundance (in log2 counts per million)
plotBCV(y.ab, cex=1)

# Fit a quasi-likelihood negative binomial generalized log-linear model to count data
fit.ab <- glmQLFit(y.ab, design, robust=TRUE, abundance.trend=FALSE)
summary(fit.ab$var.prior)

plotQLDisp(fit.ab, cex=1)

res <- glmQLFTest(fit.ab, coef=ncol(design))
summary(decideTests(res))
topTags(res)


# Visulization the results
abun.df <- reshape2::
```
